# apps/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/builds/{build_id}": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドを取得",
        "operationId": "get_build",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ビルド情報",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/builds/{build_id}/approve": {
      "post": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドを承認して baseline に昇格",
        "description": "admin 以上が必要。承認するとこのビルドの全スクリーンショットが `(project, branch)` の新しい baseline になる。未レビューの比較が残っている場合は 409。`force: true` で一括承認できる。",
        "operationId": "approve_build",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "承認後のビルド",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "承認できない状態、または未レビューの比較が残っています",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/builds/{build_id}/comparisons": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドの比較結果一覧",
        "description": "スクリーンショット名の昇順。レビュー UI が使う。",
        "operationId": "list_comparisons",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "比較結果一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonListResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/builds/{build_id}/logs": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドの進捗ログ（増分取得）",
        "description": "render / compare ジョブが追記した進捗ログを `after` カーソルで増分取得する。テナントのメンバーであること。進行中は UI がポーリングする。",
        "operationId": "get_build_logs",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "この id より後の行だけを返す（省略時は 0 = 先頭から）。",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "進捗ログの行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildLogsResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/builds/{build_id}/reject": {
      "post": {
        "tags": [
          "Builds"
        ],
        "summary": "ビルドを却下",
        "description": "admin 以上が必要。baseline は更新されず、未レビューの比較は rejected になる。",
        "operationId": "reject_build",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "却下後のビルド",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "却下できない状態です",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ci/builds/{build_id}": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "ビルドの状態を取得する（CI のポーリング用）",
        "description": "`read:build` スコープが必要。`status` が終端になるまで CI がポーリングする。",
        "operationId": "get_build_status",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "description": "ビルドID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ビルドの状態とカウント",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveBuildRequest": {
        "type": "object",
        "description": "ビルド承認リクエスト。",
        "properties": {
          "force": {
            "type": "boolean",
            "description": "`true` にすると未レビューの比較もまとめて承認する（一括承認）。"
          }
        }
      },
...
      "BuildLogEntry": {
        "type": "object",
        "description": "ビルド進捗ログの 1 行。",
        "required": [
          "id",
          "level",
          "message",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "グローバル連番。増分取得のカーソルにそのまま使う。"
          },
          "level": {
            "type": "string",
            "description": "`info` | `warn` | `error`。"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "BuildLogsResponse": {
        "type": "object",
        "description": "ビルド進捗ログの増分取得レスポンス。",
        "required": [
          "entries",
          "last_id"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuildLogEntry"
            },
            "description": "`after` より後の行（id 昇順）。"
          },
          "last_id": {
            "type": "integer",
            "format": "int64",
            "description": "クライアントが次のポーリングで `after` に渡す値。\n行が無ければリクエストされた `after` が据え置かれる。"
          }
        }
      },
...
      "ComparisonListResponse": {
        "type": "object",
        "required": [
          "comparisons",
          "total"
        ],
        "properties": {
          "comparisons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComparisonResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ServerError": {
        "type": "object",
        "description": "API 共通のエラー応答ボディ。",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "internal-error"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT"
      }
    }
  },
...
}